requirements.txt
# Zoom Manager Dependencies

# HTTP клиент для работы с API
httpx>=0.24.0

# Логирование
loguru>=0.7.0

# CLI
click>=8.0.0

# Progress bar and rich UI
rich>=13.0.0

# Environment variables
python-dotenv>=1.0.0

# База данных
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Обработка видео
ffmpeg-python>=0.2.0

# YouTube API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0

# VK API (используется aiohttp из httpx)

# Fireworks AI API
fireworks-ai>=0.15.0

# Для работы с асинхронным кодом
asyncio

# API dependencies (FastAPI, Celery, JWT)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
celery>=5.3.0
redis>=5.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
cryptography>=42.0.0
python-multipart>=0.0.6
email-validator>=2.1.0
